feat(link): support custom light clients in compiled CLI
