+++ link/cli/relayer.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-package main
+package cli
 
 import (
 	"context"
@@ -43,6 +43,7 @@ var (
 )
 
 var (
+	relayerOptions           RelayerOptions
 	flagRelayerNoMigrate     bool
 	flagRelayerHost          string
 	flagRelayerTxHash        string
@@ -55,9 +56,9 @@ func relayerRun(cmd *cobra.Command, _ []string) error {
 		return err
 	}
 
-	// nil: the CLI registers no client types beyond the built-ins. Callers
-	// embedding the relayer pass their own registry via link/app.
-	app, err := bootstrap.BuildRelayer(cfg, nil)
+	app, err := bootstrap.BuildRelayer(cfg, bootstrap.RelayerOptions{
+		LightClients: relayerOptions.LightClients,
+	})
 	if err != nil {
 		return err
 	}
