+++ link/internal/relay/proofgen/proofgen.go
@@ -12,15 +12,15 @@ import (
 
 	"github.com/cosmos/ibc/link/internal/chains"
 	"github.com/cosmos/ibc/link/internal/config"
-	pgen "github.com/cosmos/ibc/link/proofgen"
+	"github.com/cosmos/ibc/link/internal/relay/proofgen/attestation"
+	attestorservice "github.com/cosmos/ibc/link/internal/service/attestor"
+	"github.com/cosmos/ibc/link/lightclient"
 )
 
-// ProofGenerator generates packet membership/non-membership proofs and state
-// proofs for one configured light client.
-//
-// Defined in the public proofgen package so custom light clients can implement
-// it; aliased here so internal callers are unaffected.
-type ProofGenerator = pgen.ProofGenerator
+// ProofGenerator is the public light-client proof generation contract.
+// Keep the alias here so existing internal relay code can continue to refer to
+// the package that owns generator resolution and lookup.
+type ProofGenerator = lightclient.ProofGenerator
 
 // Key identifies one configured light client by the chain it lives on and
 // its client id, the composite key ProofGenerator instances are scoped by.
@@ -47,18 +47,19 @@ func (s *Set) Get(chainID, clientID string) (ProofGenerator, bool) {
 }
 
 // NewSetFromConfig resolves a ProofGenerator for every client end of every
-// configured connection, dispatching to the factory registered for each end's
-// client type.
+// configured connection. Attestation uses its built-in resolver; other client
+// types dispatch through the caller-supplied registry.
 func NewSetFromConfig(
 	ctx context.Context,
 	cfg config.Config,
 	clientSet *chains.ClientSet,
-	reg *pgen.Registry,
+	attestors []attestorservice.Attestor,
+	reg *lightclient.Registry,
 ) (*Set, error) {
 	generators := make(map[string]ProofGenerator, len(cfg.Relayer.Connections)*2)
 
 	err := forEachClientEnd(cfg, func(connAlias string, self, counterparty config.ClientEnd) error {
-		return addGenerator(ctx, generators, connAlias, self, counterparty, clientSet, reg)
+		return addGenerator(ctx, generators, connAlias, self, counterparty, clientSet, attestors, reg)
 	})
 	if err != nil {
 		return nil, err
@@ -92,8 +93,22 @@ func addGenerator(
 	connAlias string,
 	client, clientCounterparty config.ClientEnd,
 	clientSet *chains.ClientSet,
-	reg *pgen.Registry,
+	attestors []attestorservice.Attestor,
+	reg *lightclient.Registry,
 ) error {
+	if client.Type == config.ClientTypeAttestation {
+		gen, err := attestation.ResolveGenerator(
+			ctx, client, clientCounterparty, clientSet, attestors,
+		)
+		if err != nil {
+			return err
+		}
+
+		generators[Key(client.ChainID, client.ClientID)] = gen
+
+		return nil
+	}
+
 	factory, ok := reg.Get(string(client.Type))
 	if !ok {
 		return errors.Errorf(
@@ -102,19 +117,11 @@ func addGenerator(
 		)
 	}
 
-	counterpartyChain, ok := clientSet.Get(clientCounterparty.ChainID)
-	if !ok {
-		return errors.Errorf(
-			"connection %q: no configured chain client for counterparty chain %q",
-			connAlias, clientCounterparty.ChainID,
-		)
-	}
-
 	gen, err := factory.New(
 		ctx,
-		pgen.Deps{Counterparty: counterpartyChain},
-		toPublicEnd(client),
-		toPublicEnd(clientCounterparty),
+		lightclient.FactoryOptions{
+			Client: toClientInfo(client, clientCounterparty.ChainID),
+		},
 	)
 	if err != nil {
 		return errors.Wrapf(err, "connection %q", connAlias)
@@ -125,11 +132,12 @@ func addGenerator(
 	return nil
 }
 
-func toPublicEnd(end config.ClientEnd) pgen.ClientEnd {
-	return pgen.ClientEnd{
-		ChainID:  end.ChainID,
-		ClientID: end.ClientID,
-		Type:     string(end.Type),
-		Params:   end.Params,
+func toClientInfo(end config.ClientEnd, counterpartyChainID string) lightclient.ClientInfo {
+	return lightclient.ClientInfo{
+		ChainID:             end.ChainID,
+		CounterpartyChainID: counterpartyChainID,
+		ClientID:            end.ClientID,
+		Type:                string(end.Type),
+		ClientParams:        end.ClientParams,
 	}
 }
