+++ link/internal/relay/proofgen/proofgen_test.go
@@ -10,10 +10,9 @@ import (
 
 	"github.com/cosmos/ibc/link/internal/chains"
 	"github.com/cosmos/ibc/link/internal/config"
-	"github.com/cosmos/ibc/link/internal/relay/proofgen/attestation"
 	"github.com/cosmos/ibc/link/internal/service/attestor"
 	"github.com/cosmos/ibc/link/internal/tests/mocks"
-	pgen "github.com/cosmos/ibc/link/proofgen"
+	"github.com/cosmos/ibc/link/lightclient"
 )
 
 func testConnection() config.ConnectionConfig {
@@ -87,7 +86,7 @@ func TestNewSetFromConfig(t *testing.T) {
 		cfg, clientSet, attestors := testConfig(t)
 		conn := cfg.Relayer.Connections[0]
 
-		set, err := NewSetFromConfig(ctx, cfg, clientSet, testRegistry(t, clientSet, attestors))
+		set, err := NewSetFromConfig(ctx, cfg, clientSet, attestors, nil)
 		require.NoError(t, err)
 
 		_, ok := set.Get(conn.ClientA.ChainID, conn.ClientA.ClientID)
@@ -111,7 +110,7 @@ func TestNewSetFromConfig(t *testing.T) {
 		cfg := config.Config{Relayer: config.RelayerConfig{Connections: []config.ConnectionConfig{conn}}}
 
 		// ACT
-		_, err := NewSetFromConfig(ctx, cfg, clientSet, testRegistry(t, clientSet, nil))
+		_, err := NewSetFromConfig(ctx, cfg, clientSet, nil, nil)
 
 		// ASSERT
 		require.ErrorContains(t, err, `no proof generator registered for client type "tendermint"`)
@@ -132,10 +131,10 @@ func TestNewSetFromConfig(t *testing.T) {
 
 		cfg := config.Config{Relayer: config.RelayerConfig{Connections: []config.ConnectionConfig{conn}}}
 
-		reg := pgen.NewRegistry()
+		reg := lightclient.NewRegistry()
 		require.NoError(t, reg.Register("myclient", stubFactory{}))
 
-		set, err := NewSetFromConfig(ctx, cfg, clientSet, reg)
+		set, err := NewSetFromConfig(ctx, cfg, clientSet, nil, reg)
 		require.NoError(t, err)
 
 		_, ok := set.Get(conn.ClientA.ChainID, conn.ClientA.ClientID)
@@ -143,26 +142,15 @@ func TestNewSetFromConfig(t *testing.T) {
 	})
 }
 
-// testRegistry builds the built-in registry the relayer normally assembles at
-// startup.
-func testRegistry(t *testing.T, clientSet *chains.ClientSet, attestors []attestor.Attestor) *pgen.Registry {
-	t.Helper()
-
-	reg := pgen.NewRegistry()
-	require.NoError(t, reg.Register(attestation.ClientType, attestation.NewFactory(clientSet, attestors)))
-
-	return reg
-}
-
 // stubFactory is a light client type that exists only in this test.
 type stubFactory struct{}
 
-func (stubFactory) ValidateParams(*pgen.RawParams) error { return nil }
+func (stubFactory) ValidateParams(*lightclient.RawParams) error { return nil }
 
 func (stubFactory) New(
-	context.Context, pgen.Deps, pgen.ClientEnd, pgen.ClientEnd,
-) (pgen.ProofGenerator, error) {
+	context.Context, lightclient.FactoryOptions,
+) (lightclient.ProofGenerator, error) {
 	return stubGenerator{}, nil
 }
 
-type stubGenerator struct{ pgen.ProofGenerator }
+type stubGenerator struct{ lightclient.ProofGenerator }
