+++ link/internal/relay/txbuilder/evm/evm.go
@@ -123,7 +123,7 @@ func height(h uint64) ics26router.IICS02ClientMsgsHeight {
 }
 
 // packUpdateClient packs a call to updateClient(clientId, updateMsg), where
-// updateMsg is the already-encoded proof produced by proofgen.ProofGenerator.StateProof.
+// updateMsg is the already-encoded proof produced by lightclient.ProofGenerator.StateProof.
 func packUpdateClient(clientID string, updateMsg []byte) ([]byte, error) {
 	packed, err := calldata(func(opts *bind.TransactOpts) (*types.Transaction, error) {
 		return router.UpdateClient(opts, clientID, updateMsg)
